moved theme change logic to ToggleTheme comp
